+++ spark/spark_mails/apps.py
@@ -1,0 +1,7 @@
+from django.apps import AppConfig
+from django.utils.translation import ugettext_lazy as _
+
+
+class SparkMailsConfig(AppConfig):
+    name = "spark.spark_mails"
+    verbose_name = _("Spark mails")